requirements/local.txt
# Local development dependencies go here.
-r base.txt
coverage==3.6
django-cache-panel==0.1
django-discover-runner==1.0
django-debug-toolbar==0.9.4
Sphinx==1.2b1
